Weird compilation error.

+++ src/predictor/cpu_predictor.cc
@@ -254,7 +254,8 @@ class CPUPredictor : public Predictor {
                       int32_t tree_end) const {
     const int threads = omp_get_max_threads();
     constexpr double kDensityThresh = .5;
-    size_t total = std::max(p_fmat->Info().num_row_ * p_fmat->Info().num_col_, 1ul);
+    size_t total = std::max(p_fmat->Info().num_row_ * p_fmat->Info().num_col_,
+                            static_cast<uint64_t>(1));
     double density = static_cast<double>(p_fmat->Info().num_nonzero_) /
                      static_cast<double>(total);
     bool blocked = density > kDensityThresh;
